Extract extension schemas
