+++ packages/app/src/cli/services/environment.test.ts
@@ -4,7 +4,8 @@ import {selectStore, convertToTestStoreIfNeeded} from './dev/select-store.js'
 import {ensureDeploymentIdsPresence} from './environment/identifiers.js'
 import {DevEnvironmentOptions, ensureDevEnvironment, ensureDeployEnvironment, DeployAppNotFound} from './environment.js'
 import {OrganizationApp, OrganizationStore} from '../models/organization.js'
-import {App, WebType, updateAppIdentifiers, getAppIdentifiers, UIExtension} from '../models/app/app.js'
+import {App, WebType, updateAppIdentifiers, getAppIdentifiers} from '../models/app/app.js'
+import {UIExtension} from '../models/app/extensions.js'
 import {reuseDevConfigPrompt, selectOrganizationPrompt} from '../prompts/dev.js'
 import {testApp} from '../models/app/app.test-data.js'
 import {store, api, outputMocker} from '@shopify/cli-kit'
