+++ packages/app/src/cli/services/environment/id-manual-matching.test.ts
@@ -1,6 +1,6 @@
 import {manualMatchIds, ManualMatchResult} from './id-manual-matching.js'
 import {ExtensionRegistration} from '../dev/create-extension.js'
-import {UIExtension} from '../../models/app/app.js'
+import {UIExtension} from '../../models/app/extensions.js'
 import {beforeEach, describe, expect, it, vi} from 'vitest'
 import {ui} from '@shopify/cli-kit'
