+++ packages/app/src/cli/services/environment/identifiers.test.ts
@@ -3,7 +3,8 @@ import {automaticMatchmaking} from './id-matching.js'
 import {manualMatchIds} from './id-manual-matching.js'
 import {fetchAppExtensionRegistrations} from '../dev/fetch.js'
 import {createExtension, ExtensionRegistration} from '../dev/create-extension.js'
-import {App, FunctionExtension, UIExtension} from '../../models/app/app.js'
+import {App} from '../../models/app/app.js'
+import {FunctionExtension, UIExtension} from '../../models/app/extensions.js'
 import {beforeEach, describe, expect, it, vi} from 'vitest'
 import {ui} from '@shopify/cli-kit'
