+++ packages/app/src/cli/services/environment/identifiers.ts
@@ -1,6 +1,7 @@
 import {automaticMatchmaking} from './id-matching.js'
 import {manualMatchIds} from './id-manual-matching.js'
-import {App, Extension, Identifiers} from '../../models/app/app.js'
+import {App, Identifiers} from '../../models/app/app.js'
+import {Extension} from '../../models/app/extensions.js'
 import {fetchAppExtensionRegistrations} from '../dev/fetch.js'
 import {createExtension} from '../dev/create-extension.js'
 import {dependency, error, output, session, ui} from '@shopify/cli-kit'
